Fuse fmpz_mpoly/set* apart from set_str_pretty
